Merge Sorted Array
Intuition
The intuition behind this solution is to be a backward merging approach. Starting from the end of nums1 (which has enough space to accommodate the merged elements), it compares the elements from the end of nums1 and nums2 and places the larger element at the end of nums1. This process continues until either m or n becomes 0. If there are remaining elements in nums2 after this loop, they are copied to the beginning of nums1.

Approach
Initialize size to be the total length of the merged array (m + n).
Use a while loop to iterate while both m and n are greater than 0.
If the element at nums1[m - 1] is less than or equal to the element at nums2[n - 1], place the element at nums2[n - 1] at the end of nums1 (at position size - 1). Decrement n.
If the element at nums1[m - 1] is greater than or equal to the element at nums2[n - 1], place the element at nums1[m - 1] at the end of nums1. Decrement m.
Decrement size in each iteration.
If m becomes 0 but there are remaining elements in nums2, copy those remaining elements to the beginning of nums1.
Complexity
Time complexity:
The while loop iterates until either m or n becomes 0. In each iteration, a constant amount of work is done.

The time complexity is O(m + n), which is generally more efficient than the sorting-based solution.

Space complexity:
The space complexity is O(1) as no extra space is used; the modifications are done in-place.