Remove Element
Intuition
The given problem requires removing all occurrences of a specific value (val) from the array nums in-place. Additionally, the order of the elements may change, and we need to return the count of elements that are not equal to val.

Approach
The provided solution uses two pointers, s and e, initialized at the start and end of the array, respectively. It then iterates through the array, swapping elements at the pointers if needed to move occurrences of val to the end of the array.

Initialize s and e at the start and end of the array, i.e., s = 0 and e = len(nums) - 1.
Use a while loop to iterate through the array.
If nums[s] is equal to val and nums[e] is not equal to val, swap nums[s] and nums[e], then increment s and decrement e.
If nums[e] is equal to val, decrement e.
If neither of the above conditions is met, increment s.
Continue this process until s is greater than e.
The algorithm modifies the array in-place, and the final count of elements not equal to val is the value of s.
Complexity
Time complexity:
O(n) where n is the length of the array. The algorithm processes each element at most once.

Space complexity:
O(1) The algorithm uses only a constant amount of extra space regardless of the size of the input array.