Contains Duplicate leetcode
Intuition
The intuition behind this solution is to use a hash set to keep track of unique elements as we iterate through the input array. If at any point we encounter an element that already exists in the hash set, it means we have found a duplicate, and we can return True. If we iterate through the entire array without finding any duplicates, we return False.

Approach
Create an empty hash set called hashset to store unique elements.
Iterate through the elements of the input array nums.
For each element i:
Check if i is already in the hash set.
If it is, return True as we have found a duplicate.
If it's not, add i to the hash set.
If the loop completes without finding any duplicates, return False.
Complexity
Time complexity:
O(n). In the worst case, we iterate through all elements in the array once.
Space complexity:
O(n). The hash set may store all elements in the array in the worst case, leading to linear space complexity.