Valid Anagram
Intuition
The intuition behind this solution is to use two dictionaries (count_s and count_t) to store the frequency of characters in strings s and t. Then, compare these dictionaries to check if the two strings are anagrams.

Approach
Check if the lengths of strings s and t are equal. If not, return False as they cannot be anagrams.
Create two dictionaries (count_s and count_t) to store the frequency of characters in strings s and t.
Iterate through each character in both strings and update the respective dictionaries.
After iterating through both strings, compare the two dictionaries. If they are equal, return True; otherwise, return False.
Complexity
Time complexity:
O(n), where n is the length of the input strings s and t. The function iterates through each character in both strings once.
Space complexity:
O(c), where c is the number of unique characters in the input strings s and t. The function uses two dictionaries to store the frequency of characters. In the worst case, if all characters are unique, the space complexity is proportional to the number of unique characters.